document_parser.py
#!/usr/bin/env python3
import os
import magic
import io

class DocumentParser:
    """
    Parser to extract text from various document types.
    
    Note: This is a placeholder implementation. In a real implementation,
    you would need to install and import additional libraries:
    - PyPDF2 or pdfplumber for PDFs
    - python-docx for DOCX files
    - pandas for Excel files
    etc.
    """
    
    def __init__(self):
        """Initialize the document parser."""
        self.parsers = {
            'text/plain': self._parse_text,
            'application/pdf': self._parse_pdf,
            'application/msword': self._parse_doc,
            'application/vnd.openxmlformats-officedocument.wordprocessingml.document': self._parse_docx,
            'application/vnd.ms-excel': self._parse_excel,
            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet': self._parse_xlsx,
            'text/csv': self._parse_csv,
            'text/rtf': self._parse_rtf,
        }
    
    def parse(self, file_path):
        """
        Parse a document file and extract text.
        
        Args:
            file_path: Path to the document file
            
        Returns:
            str: Extracted text from the document
        """
        try:
            # Determine file type
            mime_type = magic.from_file(file_path, mime=True)
            
            # Get the appropriate parser
            parser = self.parsers.get(mime_type, self._parse_unknown)
            
            # Parse the document
            return parser(file_path)
        
        except Exception as e:
            print(f"Error parsing document {file_path}: {str(e)}")
            return ""
    
    def _parse_text(self, file_path):
        """Parse a plain text file."""
        try:
            with open(file_path, 'r', errors='ignore') as file:
                return file.read()
        except Exception as e:
            print(f"Error parsing text file {file_path}: {str(e)}")
            return ""
    
    def _parse_pdf(self, file_path):
        """
        Parse a PDF file.
        
        In a real implementation, you would use PyPDF2 or pdfplumber:
        
        ```python
        import PyPDF2
        
        with open(file_path, 'rb') as file:
            reader = PyPDF2.PdfFileReader(file)
            text = ""
            for page_num in range(reader.numPages):
                text += reader.getPage(page_num).extractText()
            return text
        ```
        """
        return f"[Placeholder: PDF text extraction from {file_path}. Install PyPDF2 for real implementation]"
    
    def _parse_doc(self, file_path):
        """
        Parse a DOC file.
        
        In a real implementation, you would use a library that can handle DOC files.
        """
        return f"[Placeholder: DOC text extraction from {file_path}. Install appropriate libraries for real implementation]"
    
    def _parse_docx(self, file_path):
        """
        Parse a DOCX file.
        
        In a real implementation, you would use python-docx:
        
        ```python
        import docx
        
        doc = docx.Document(file_path)
        text = "\n".join([paragraph.text for paragraph in doc.paragraphs])
        return text
        ```
        """
        return f"[Placeholder: DOCX text extraction from {file_path}. Install python-docx for real implementation]"
    
    def _parse_excel(self, file_path):
        """
        Parse an XLS file.
        
        In a real implementation, you would use pandas or xlrd.
        """
        return f"[Placeholder: XLS text extraction from {file_path}. Install pandas or xlrd for real implementation]"
    
    def _parse_xlsx(self, file_path):
        """
        Parse an XLSX file.
        
        In a real implementation, you would use pandas or openpyxl:
        
        ```python
        import pandas as pd
        
        # Read all sheets
        xls = pd.ExcelFile(file_path)
        text = ""
        for sheet_name in xls.sheet_names:
            df = pd.read_excel(xls, sheet_name)
            text += f"Sheet: {sheet_name}\n"
            text += df.to_string() + "\n\n"
        return text
        ```
        """
        return f"[Placeholder: XLSX text extraction from {file_path}. Install pandas or openpyxl for real implementation]"
    
    def _parse_csv(self, file_path):
        """
        Parse a CSV file.
        
        In a real implementation, you would use pandas or csv module:
        
        ```python
        import pandas as pd
        
        df = pd.read_csv(file_path)
        return df.to_string()
        ```
        """
        return f"[Placeholder: CSV text extraction from {file_path}. Install pandas for real implementation]"
    
    def _parse_rtf(self, file_path):
        """
        Parse an RTF file.
        
        In a real implementation, you would use a library that can handle RTF files.
        """
        return f"[Placeholder: RTF text extraction from {file_path}. Install appropriate libraries for real implementation]"
    
    def _parse_unknown(self, file_path):
        """Handle unknown file types."""
        return f"[Unable to extract text from unknown file type: {file_path}]" 