requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-magic==0.4.27
google-generativeai==0.3.1
python-dotenv==1.0.0
tqdm==4.66.1
PySocks==1.7.1
stem==1.8.1 